% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XX}
\alias{XX}
\title{My data}
\format{
\subsection{\code{XX}}{

A data frame with XX rows and XX columns:
\describe{
\item{lines}{lines of text}
\item{author}{who wrote that line}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
XX
}
\description{
This is more descriptive text about my data.
Even more info is provided here ...
}
\keyword{datasets}
